fix and feat: the pomodoroTimer class was separated from the view and some bugs were fixed
