+++ lib/pages/dialog_helper.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:tempo_app/model/model_pomodoro.dart';
+import 'package:tempo_app/model/model_pomodoro_timer.dart';
+import 'package:tempo_app/pages/view_pomodoro.dart';
 import 'package:tempo_app/pages/view_tasks.dart';
 
 class DialogHelper {
@@ -38,6 +40,7 @@ class DialogHelper {
                 Navigator.of(context).push(
                     MaterialPageRoute(builder: (context) => const Tasks()));
                 Pomodoro.actualPomodoro = null;
+                PomodoroTimer.currentSessionIndex = 0;
               },
               child: const Text('Cerrar'),
             ),
